requirements.txt
Flask==2.2.5
psycopg2-binary>=2.9.9
SQLAlchemy==1.4.46
Flask-Migrate==4.0.4
python-dotenv==1.0.0
gunicorn==22.0.0
redis==5.0.1
Flask-Session==0.8.0
python-json-logger==2.0.7